+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # chat21-web-widget ver 2.0
 
+### 1.010
+- Fixes bug load image profile 
+- limit number agents to 5
+- change css ballon agent
+- close callout on openwidget
+- add external params showWaitTime, showAvailableAgents, showLogoutOption;
+
 ### 1.009
 - Fixes align default to right